+++ Misc/NEWS.d/next/C API/2021-02-17-18-51-26.bpo-35134.YoQdk8.rst
@@ -1,0 +1,3 @@
+Move pyarena.h, pyctype.h, and pytime.h into the cpython/ directory. They
+must not be included directly, as they are already included by Python.h:
+:ref:`Include Files <api-includes>`.
